% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{aggregateFeatureMatrix}
\alias{aggregateFeatureMatrix}
\title{This function takes a matrix where rows are features and columns are cells,
and a neighbourhood list, and creates an matrix where columns are the
neighbourhoods, the rows are are the features and the values are aggregated
expression values for cells in each neighbourhood.}
\usage{
aggregateFeatureMatrix(M, nbhdList, aggregateFunction)
}
\arguments{
\item{M}{\itemize{
\item a matrix where column names are cells and row names are
features.
}}

\item{nbhdList}{\itemize{
\item a named list with memberships of the neighbourhoods
of cells
}}

\item{aggregateFunction}{\itemize{
\item a function to aggregate expression (e.g. rowSums,
rowMeans)
}}
}
\value{
a matrix giving aggregated gene expression for a cell's neighbourhood.
}
\description{
This function takes a matrix where rows are features and columns are cells,
and a neighbourhood list, and creates an matrix where columns are the
neighbourhoods, the rows are are the features and the values are aggregated
expression values for cells in each neighbourhood.
}
\examples{
aggregateFeatureMatrix = function(M, nbhdList, aggregateFunction)
}
