% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleGeneAnnotation.R
\name{annotateGenesByGeneSet}
\alias{annotateGenesByGeneSet}
\title{This function annotates genes with terms}
\usage{
annotateGenesByGeneSet(geneSets)
}
\arguments{
\item{geneSets}{\itemize{
\item a list of gene sets, e.g., as produced by
readGmt
}}
}
\value{
\itemize{
\item A list where names are genes and values are lists of terms
}
}
\description{
This essentially inverts a list of gene sets.  It takes
a list (e.g., Hallmark or GO) where each list item is a
name of a gene set and gives the genes in that set and
returns a list where each item is a gene and gives the
gene sets that gene is in.
}
\examples{
hallmark = make.getExample()('hallmark')
annotatedGenes = annotateGenesByGeneSet(hallmark)
}
