% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{annotateInteractionCounts}
\alias{annotateInteractionCounts}
\title{This takes a data frame of interaction counts as found
by countInteractionsPerCell(), the underlying Seurat object
and the neighbourhood Seurat object and annotates the counts
with the cell type and the neighbourhood type corresponding
to the cells of the interaction counts.}
\usage{
annotateInteractionCounts(interactionCounts, obj, nbhdObj)
}
\arguments{
\item{interactionCounts}{\itemize{
\item as found by countInteractionsPerCell()
}}

\item{obj}{\itemize{
\item a Seurat object
}}

\item{nbhdObj}{\itemize{
\item a neighbourhood x cell type Seurat object
}}
}
\value{
This returns the interaction counts annotated with the
cell type and neighbourhood type of each cell.
}
\description{
This takes a data frame of interaction counts as found
by countInteractionsPerCell(), the underlying Seurat object
and the neighbourhood Seurat object and annotates the counts
with the cell type and the neighbourhood type corresponding
to the cells of the interaction counts.
}
