% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{annotateLRInteractionCounts}
\alias{annotateLRInteractionCounts}
\title{This takes a data frame of interaction counts as found
by countLRInteractionsPerCell(), the underlying Seurat object
and the neighbourhood Seurat object and annotates the counts
with the cell type and the neighbourhood type corresponding
to the cells of the interaction counts.}
\usage{
annotateLRInteractionCounts(interactionCounts, obj, nbhdObj)
}
\arguments{
\item{interactionCounts}{\itemize{
\item as found by countLRInteractionsPerCell()
}}

\item{obj}{\itemize{
\item a Seurat object, or SingleCellExperiment to be turned
into a Seurat object
}}

\item{nbhdObj}{\itemize{
\item a neighbourhood x cell type Seurat object or a
SingleCellExperiment to be turned into a Seurat object
}}
}
\value{
This returns the interaction counts annotated with the
cell type and neighbourhood type of each cell.
}
\description{
This takes a data frame of interaction counts as found
by countLRInteractionsPerCell(), the underlying Seurat object
and the neighbourhood Seurat object and annotates the counts
with the cell type and the neighbourhood type corresponding
to the cells of the interaction counts.
}
