% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{averageExpMatrix}
\alias{averageExpMatrix}
\title{averageExpMatrix}
\format{
A 30 x 13 matrix
\describe{
This gives the average gene expression for each of the
gene clusters in each of the cell clusters.
}
}
\source{
Produced from S and STranspose by
getAverageExpression()
}
\usage{
averageExpMatrix
}
\description{
The average expression matrix for cell clusters of S
and gene clusters of STranspose.
}
\keyword{datasets}
