% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{cellTypesPerCellTypeMatrix}
\alias{cellTypesPerCellTypeMatrix}
\title{For each cell type, this function looks at the neighbourhoods
around cells of that type and discovers the fractions of those
cells of each type.}
\usage{
cellTypesPerCellTypeMatrix(nbhdByCellType, seurat_clusters)
}
\arguments{
\item{nbhdByCellType}{\itemize{
\item A matrix whose rows are neighbourhoods
each denoted by the cell at their center, whose columns are
cell types, and whose entries are counts.
}}

\item{seurat_clusters}{\itemize{
\item a named vector whose names are the cells
and whose entries are their seurat_clusters.
}}
}
\value{
A square matrix whose rownames and colnames are the
seurat_clusters as character strings.  Each row corresponds
to neighbourhoods around all cells of that type and the entries
give the fractions of those neighbourhoods occupied by cells
of each type.
}
\description{
For each cell type, this function looks at the neighbourhoods
around cells of that type and discovers the fractions of those
cells of each type.
}
