% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids}
\alias{centroids}
\title{centroids}
\format{
a dataframe
where rownames are cellnames and columns contain x
and y coordinates respectively.
\describe{
This gives the x and y coordinates for cell centroids from xenium
mouse brain spatial data.
}
}
\source{
subset of tiny subset from
https://www.10xgenomics.com/resources/datasets/fresh-frozen-mouse-brain-for-xenium-explorer-demo-1-standard
}
\usage{
centroids
}
\description{
Cell centroids from xenium spatial data
}
\keyword{datasets}
