% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusterDF}
\alias{clusterDF}
\title{clusterDF}
\format{
A data frame with columns gene and geneCluster
\describe{
This gives the cluster membership of the genes of STranspose.
}
}
\source{
Extracted from STranspose
}
\usage{
clusterDF
}
\description{
A data frame of genes and their gene cluster membership
}
\keyword{datasets}
