% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clusters}
\alias{clusters}
\title{clusters}
\format{
a named vector of cluster where names are each cell and
clusters are a factor.
\describe{
This contains cluster annotations for xenium mouse brain data extracted from
a Seurat analysis.
}
}
\source{
clusters from Seurat analysis of
https://www.10xgenomics.com/resources/datasets/fresh-frozen-mouse-brain-for-xenium-explorer-demo-1-standard
}
\usage{
clusters
}
\description{
Clusters from xenium spatial data
}
\keyword{datasets}
