% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{collapseNeighbourhoods}
\alias{collapseNeighbourhoods}
\title{This function takes expanded neighbourhoods and collapses them to a nearest
neighbourhood graph where all neighbours of degree <= n in the original graph
are considered first neighbours.}
\usage{
collapseNeighbourhoods(expandedNeighbours, n = length(expandedNeighbours))
}
\arguments{
\item{expandedNeighbours}{\itemize{
\item the results of expandNeighbourhoods()
}}

\item{n}{\itemize{
\item the maximum degree to connect neighbours. Defaults to the maximum
degree neighbourhoods were expanded to in the results of expandNeighbourhoods().
}}
}
\value{
a graph in neighbour format, i.e., a data frame with
columns nodeA and nodeB, where nodes that were originally of degree <= n
are connected.
}
\description{
This function takes expanded neighbourhoods and collapses them to a nearest
neighbourhood graph where all neighbours of degree <= n in the original graph
are considered first neighbours.
}
