% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colours}
\alias{colours}
\title{colours}
\format{
A character vector of colours where names are clusters.

\describe{
A character vector of colours from the polychrome palette where names
are clusters.
}
}
\source{
Created using Seurat DiscretePalette()
}
\usage{
colours
}
\description{
A character vector of colours where names are clusters.
}
\keyword{datasets}
