% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeGraphEmbedding}
\alias{computeGraphEmbedding}
\title{This function adds a force directed graph embedding to a seurat object}
\usage{
computeGraphEmbedding(seuratObj, graph = defaultGraph(seuratObj))
}
\arguments{
\item{seuratObj}{\itemize{
\item a seurat object.
}}

\item{graph}{\itemize{
\item which graph to extract.  Defaults to
paste0(f@active.assay,'_snn')
}}
}
\value{
a seurat object with a "graph" dimensionality reduction.
}
\description{
This function adds a force directed graph embedding to a seurat object
}
\examples{
objWithEmbedding = computeGraphEmbedding(NBHDByCTSeurat)
}
