% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleNeighbourhoodAnalysis.R
\name{computeMoransI}
\alias{computeMoransI}
\title{This function takes a matrix where rows are features and columns are cells,
and a neighbourhood list, and computes Moran's I.}
\usage{
computeMoransI(M, nbhdList)
}
\arguments{
\item{M}{\itemize{
\item a matrix where column names are cells and row names are features.
}}

\item{nbhdList}{\itemize{
\item a named list with memberships of the neighbourhoods
of cells
}}
}
\value{
a matrix giving aggregated gene expression for a cell's neighbourhood.
}
\description{
This function takes a matrix where rows are features and columns are cells,
and a neighbourhood list, and computes Moran's I.
}
