% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeNBHDVsCTSeurat}
\alias{computeNBHDVsCTSeurat}
\title{This function creates a seurat object using a neighbourhood by cell type
matrix}
\usage{
computeNBHDVsCTSeurat(
  dataMatrix,
  resolution = 0.1,
  npcs = 10,
  n.neighbors = 30L,
  transpose = F,
  verbose = T
)
}
\arguments{
\item{dataMatrix}{\itemize{
\item a matrix of neighbourhoods by cell types or its
transpose.
}}

\item{resolution}{\itemize{
\item resolution for clustering (default 0.1).
}}

\item{npcs}{\itemize{
\item number of pcs used for PCA, defaults to 10.
}}

\item{n.neighbors}{\itemize{
\item number of neighbors used by UMAP, defaults to 30.
}}

\item{transpose}{\itemize{
\item defaults to FALSE.
}}

\item{verbose}{\itemize{
\item defaults to TRUE, used to limit trace if FALSE
}}
}
\value{
a seurat object based on a neighbourhood by cell type matrix or its
transpose, containing clusters and UMAP.
}
\description{
This function creates a seurat object using a neighbourhood by cell type
matrix
}
\examples{
NBHDByCTSeurat = computeNBHDVsCTSeurat(NBHDByCTMatrix)
}
