% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeNeighbourhoodByCTSeurat}
\alias{computeNeighbourhoodByCTSeurat}
\title{This function creates a seurat object using a neighbourhood by cell type
matrix}
\usage{
computeNeighbourhoodByCTSeurat(
  neighbourhoodByCT,
  resolution = 0.1,
  npcs = 10,
  n.neighbors = 30L,
  compute_UMAP = T,
  transpose = F
)
}
\arguments{
\item{neighbourhoodByCT}{\itemize{
\item a matrix of neighbourhoods by cell types
}}

\item{resolution}{\itemize{
\item resolution for clustering (default 0.1)
}}
}
\value{
a seurat object based on a neighbourhood by cell type matrix,
containing clusters and UMAP.
}
\description{
This function creates a seurat object using a neighbourhood by cell type
matrix
}
