% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeNeighboursEuclidean}
\alias{computeNeighboursEuclidean}
\title{This function computes a spatial graph where
neighbors are identified based on euclidean distance and a
user defined threshold.}
\usage{
computeNeighboursEuclidean(coords, threshold)
}
\arguments{
\item{threshold}{\itemize{
\item a distance cut off to compute neighbours.
}}

\item{centroids}{\itemize{
\item a dataframe containing centroids
where rownames are cellnames and columns contain x
and y coordinates respectively.
}}
}
\value{
a graph in neighbour format, i.e., a data frame with
columns nodeA and nodeB.
}
\description{
This function computes a spatial graph where
neighbors are identified based on euclidean distance and a
user defined threshold.
}
