% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleLigandReceptorAnalysis.R
\name{convertToLong}
\alias{convertToLong}
\title{This is a utility function for converting ligandReceptor cluster-level
results to long format and calculates adjusted p-values.}
\usage{
convertToLong(ligandReceptorResults)
}
\arguments{
\item{ligandReceptorResults}{\itemize{
\item ligandReceptorReults calculated using
performLigandReceptorAnalysis()
}}
}
\value{
ligand receptor results in long format
}
\description{
This is a utility function for converting ligandReceptor cluster-level
results to long format and calculates adjusted p-values.
}
