% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{countInteractionsPerCell}
\alias{countInteractionsPerCell}
\title{This function takes a listing of the neighbouring
cells together with the presence or absence of each
ligand-receptor pair on each edge and produces a count
showing for each cell, how many neighbours it has with
that interaction either as source or as target}
\usage{
countInteractionsPerCell(edges, sourceOrTarget)
}
\arguments{
\item{edges}{\itemize{
\item A data frame of neighbouring cells
together with their interactions as produced by
getInteractionsOnEdges()
}}

\item{sourceOrTarget}{\itemize{
\item a character, either 'source' or
'target' telling which direction of interaction to count
}}
}
\value{
This returns a data frame with one row for each
cell and a column giving the name of that cell and the
other columns giving the counts of interactions that it
has with its neighbours.
}
\description{
This function takes a listing of the neighbouring
cells together with the presence or absence of each
ligand-receptor pair on each edge and produces a count
showing for each cell, how many neighbours it has with
that interaction either as source or as target
}
