% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CTByNBHDSeurat}
\alias{CTByNBHDSeurat}
\title{CTByNBHDSeurat}
\format{
A Seurat object consisting of 24 samples (the cell types) and
4261 features (the neighbourhoods).

\describe{
This is a Seurat object created by taking t(NBHDByCTMatrix)
as the counts.
}
}
\source{
Created from t(NBHDByCTMatrix) by
computeNBHDVsCTSeurat()
}
\usage{
CTByNBHDSeurat
}
\description{
A Seurat object computed from the transpose of NBHDByCTMatrix. Think of
cell types "expressing" (being found in) neighbourhoods.
}
\keyword{datasets}
