% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{delaunayNeighbours}
\alias{delaunayNeighbours}
\title{delaunayNeighbours}
\format{
A data frame with character columns nodeA and nodeB.

\describe{
The entries in this data frame are cell names from smallXenium.
Each row represents an undirected edge in the Delaunay
triangulation of centroids.
}
}
\source{
This is computed from centroids using
computeNeighboursDelaunay()
}
\usage{
delaunayNeighbours
}
\description{
A data frame of the Delaunay triangulation of centroids of cells in
smallXenium
}
\keyword{datasets}
