% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{edgeCutoffsByClustering}
\alias{edgeCutoffsByClustering}
\title{This finds proposed cutoffs for edge lengths by clustering
the lengths of the edges for each cell type pair using k-means
clustering with k  = 2}
\usage{
edgeCutoffsByClustering(annEdges)
}
\value{
This returns a data frame with columns cellTypePair and
cutoff.
}
\description{
@param annEdges - a data frame with columns nodeA, nodeB, length
and cellPairType as produced by edgeLengthsAndCellPairTypes.
}
\examples{
annEdges =
    edgeLengthsAndCellPairTypes(delaunayNeighbours,clusters,centroids)
cutoffDF = edgeCutoffsByClustering(annEdges)
}
