% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgeNeighbours}
\alias{edgeNeighbours}
\title{edgeNeighbours}
\format{
A data frame with character columns nodeA and nodeB.

\describe{
A spatial graph where edges in the original delaunayNeighbours become nodes
and A-B edges (in the original graph) become connected to
all A- edges and all B- edges.
}
}
\source{
This is computed from delaunayNeighbours using
computeEdgeGraph()
}
\usage{
edgeNeighbours
}
\description{
A neighbour graph where nodes represent interactions between cells.
}
\keyword{datasets}
