% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edgeSeurat}
\alias{edgeSeurat}
\title{edgeSeurat}
\format{
A Seurat object consisting of 25518 samples (edges between cells
and 28 features (ligand-receptor pairs).

\describe{
This is a Seurat object where
each point represents an edge between cells, and spatial coordinates are the
centroids of edges between cells. The "expression matrix" is the
binarised presence/absence of an interaction (ligand receptor pair) on an edge.
}
}
\source{
Created from ligandReceptorResults by
computeEdgeSeurat()
}
\usage{
edgeSeurat
}
\description{
A Seurat object computed from ligandReceptorResults using
computeEdgeSeurat()
}
\keyword{datasets}
