% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{euclideanNeighbours}
\alias{euclideanNeighbours}
\title{euclideanNeighbours}
\format{
A data frame with character columns nodeA and nodeB.

\describe{
The entries in this data frame are cell names from smallXenium.
Each row represents an undirected edge in the nearest neighbour
graph using a cutoff euclidean distance of 20 \mu m.
}
}
\source{
This is computed from centroids using
computeNeighboursEuclidean()
}
\usage{
euclideanNeighbours
}
\description{
A data frame of nearest neighbours calculated
using euclidean distance as a cutoff.
}
\keyword{datasets}
