% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleExampleServer.R
\name{exampleObjects}
\alias{exampleObjects}
\title{This returns the names of available example objects.}
\usage{
exampleObjects()
}
\value{
A character vector of the names of available
example data objects
}
\description{
This returns the names of available example objects.
}
\examples{
availableObjects = exampleObjects()
}
