% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{expandNeighbourhoods}
\alias{expandNeighbourhoods}
\title{This function takes a nearest neighbour graph and a radius
and calculates nth neighbour graphs where max(n) == radius}
\usage{
expandNeighbourhoods(spatialGraph, n)
}
\arguments{
\item{spatialGraph}{\itemize{
\item a nearest neighbour graph
}}

\item{n}{\itemize{
\item the maximum degree to calculate a neighbour graph with edges
connecting vertices of degree n for.
}}
}
\value{
A named list of neighbour graphs, where each graph contains edges
connecting vertices of degree n. Each graph is named according to degree n.
}
\description{
This function takes a nearest neighbour graph and a radius
and calculates nth neighbour graphs where max(n) == radius
}
\examples{
smallNbhds = expandNeighbourhoods(smallDelaunayTriangulation,4)
}
