% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extendedNeighbours}
\alias{extendedNeighbours}
\title{extendedNeighbours}
\format{
A data frame with character columns nodeA and nodeB.

\describe{
The entries in this data frame are cell names from smallXenium.
Each row represents an undirected edge in the extended neighbour
graph. Nodes are connected to all nodes up to
and including degree 4 in the original graph (delaunayNeighbours).
}
}
\source{
This is computed from extendedNeighboursList using
collapseExtendedNBHDs()
}
\usage{
extendedNeighbours
}
\description{
A neighbour graph where nodes are connected to all nodes up to
and including degree 4 in the original graph (delaunayNeighbours)
}
\keyword{datasets}
