% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extendedNeighboursList}
\alias{extendedNeighboursList}
\title{extendedNeighboursList}
\format{
A list of dataframes with character columns nodeA and nodeB.

\describe{
A named list of neighbour graphs, where each graph contains edges
connecting vertices of degree n. Each graph (list entry) is named according to degree n.
}
}
\source{
This is computed from delaunayNeighbours using
getExtendedNBHDs()
}
\usage{
extendedNeighboursList
}
\description{
A list of nth degree neighbour graphs calculated from delaunayNeighbours.
}
\keyword{datasets}
