% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{findCombinatorialNeighbourhoods}
\alias{findCombinatorialNeighbourhoods}
\title{This function takes a nearest neighbour graph and a radius
and finds the combinatorial ball around each cell}
\usage{
findCombinatorialNeighbourhoods(NN, radius)
}
\arguments{
\item{NN}{\itemize{
\item a nearest neighbour graph
}}

\item{radius}{\itemize{
\item combinatorial radius
}}
}
\value{
This returns a named list.  In each case, the name is the
cell at the center of the ball.  Each item in the list is a
named numeric.  The values give combinatorial distance from the
center and the names give the cells.
}
\description{
This function takes a nearest neighbour graph and a radius
and finds the combinatorial ball around each cell
}
