% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleLigandReceptorAnalysis.R
\name{formatData}
\alias{formatData}
\title{This is a utility function for converting entries in ligandReceptorResults
to long format.}
\usage{
formatData(data, name)
}
\arguments{
\item{data}{\itemize{
\item item from ligandReceptorResults
}}

\item{name}{\itemize{
\item name to give column of returned data
}}
}
\value{
dataframe with item from ligandReceptorResults in long format
}
\description{
This is a utility function for converting entries in ligandReceptorResults
to long format.
}
