% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{geneSetsVsGeneClustersPValueMatrix}
\alias{geneSetsVsGeneClustersPValueMatrix}
\title{This compares the gene clusters to other gene sets
e.g., GO, Hallmark, and determines the p-value for
their overlaps when compared to a set of background
genes.}
\usage{
geneSetsVsGeneClustersPValueMatrix(geneSets, clusterDF, backgroundGenes)
}
\arguments{
\item{geneSets}{\itemize{
\item a named list of gene sets
}}

\item{clusterDF}{\itemize{
\item a data frame giving the cluster
membership of each gene with columns gene and geneCluster
}}

\item{backgroundGenes}{\itemize{
\item a character vector of genes
}}
}
\description{
This compares the gene clusters to other gene sets
e.g., GO, Hallmark, and determines the p-value for
their overlaps when compared to a set of background
genes.
}
