% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{getAverageExpressionDF}
\alias{getAverageExpressionDF}
\title{This converts an average gene expression matrix to a
data frame.}
\usage{
getAverageExpressionDF(M)
}
\arguments{
\item{M}{\itemize{
\item An average gene expression matrix.
}}
}
\value{
A data frame with columns cellCluster, geneCluster
and average expression
}
\description{
This converts an average gene expression matrix to a
data frame.
}
