% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleGenesVsCells.R
\name{getFeatureZScores}
\alias{getFeatureZScores}
\title{This gets z-scores for the values of features}
\usage{
getFeatureZScores(f, features = rownames(f), layer = "data")
}
\arguments{
\item{f}{\itemize{
\item a Seurat object of cells or SingleCellExperiment to
be converted to a Seurat object
}}

\item{features}{\itemize{
\item a set of features to retrieve z-scores for,
defaults to rownames(f)
}}

\item{layer}{\itemize{
\item the data layer to retrieve
}}
}
\value{
This returns a data frame with a column for each
feature and a row for each cell
}
\description{
This gets z-scores for the values of features
}
\examples{
getExample = make.getExample()
exSeuratObj = getExample('exSeuratObj',toy=TRUE)
df = getFeatureZScores(exSeuratObj)
}
