% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{getGeneNeighbors}
\alias{getGeneNeighbors}
\title{This function gets the neighbors of a given gene
using either the gene Seurat object or its nearest
neighbor graph returned from getNearestNeighborListsSeurat}
\usage{
getGeneNeighbors(gene, NN)
}
\arguments{
\item{gene}{\itemize{
\item the gene in question
}}

\item{NN}{\itemize{
\item either the gene Seurat object or its nearest
neighbor graph as found by getNearestNeighborListsSeurat
}}
}
\value{
the neighboring genes
}
\description{
This function gets the neighbors of a given gene
using either the gene Seurat object or its nearest
neighbor graph returned from getNearestNeighborListsSeurat
}
