% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleLigandReceptorAnalysis.R
\name{getLigandReceptorNetwork}
\alias{getLigandReceptorNetwork}
\title{This function retrieves the Nichenetr ligand-
receptor network for mouse or human.}
\usage{
getLigandReceptorNetwork(species)
}
\arguments{
\item{species}{\itemize{
\item either 'human' or 'mouse'
}}
}
\value{
This returns a data frame whose first two
columns are from and to, i.e., ligand and receptor.
These are derived from the nichenetr ligand receptor
networks.
}
\description{
This function retrieves the Nichenetr ligand-
receptor network for mouse or human.
}
\examples{
lrn = getLigandReceptorNetwork('human')
}
