% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleGeometricClustering.R
\name{getSubsetComponents}
\alias{getSubsetComponents}
\title{This is designed to dectect the components of a
gene subset in the case where median complement
distance detects clustering.}
\usage{
getSubsetComponents(fPrime, theSubset, alpha = 0.5, edgeCut = NA)
}
\arguments{
\item{fPrime}{\itemize{
\item a gene Seurat object or
SingleCellExperiment
}}

\item{theSubset}{\itemize{
\item a subset of the genes
}}

\item{alpha}{\itemize{
\item a parameter typically less than
one controling the granularity of the components.
Defaults to .5
}}

\item{edgeCut}{\itemize{
\item the maximum length of edges included in
the subgraph whose components are returned.  If it
is NA (the default) it is computed using alpha.
Otherwise, it can be supplied directly.
}}
}
\value{
A list of the components of the subset
treated as a graph whose edges are determined by
their distance in UMAP coordinates.
}
\description{
This is designed to dectect the components of a
gene subset in the case where median complement
distance detects clustering.
}
