% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hallmark}
\alias{hallmark}
\title{hallmark}
\format{
A named list
\describe{
The mouse hallmark gene sets.
}
}
\source{
https://www.gsea-msigdb.org/gsea/msigdb/mouse/collections.jsp
}
\usage{
hallmark
}
\description{
The mouse hallmark gene sets.
}
\keyword{datasets}
