% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanLNR}
\alias{humanLNR}
\title{humanLRN}
\format{
a data frame with two columns, 'from' and 'to'
\describe{A data frame with two columns, 'from' and 'to'.
Each row represents a human ligand - receptor pair.
}
}
\source{
This is derived from the nichenetr human ligand -
receptor network.
}
\usage{
humanLNR
}
\description{
A data frame giving 12019 human ligand receptor pairs
}
\keyword{datasets}
