% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{humanLRN}
\alias{humanLRN}
\title{humanLRN}
\format{
a data frame with two columns, 'from' and 'to'
\describe{A data frame with two columns, 'from' and 'to'.
Each row represents a human ligand - receptor pair.
}
}
\source{
This is taken from the nichenetr package,
url = {https://www.nature.com/articles/s41592-019-0667-5}.
Specifically we use the human ligand - receptor network.
}
\usage{
humanLRN
}
\description{
A data frame giving 12019 human ligand receptor pairs
}
\keyword{datasets}
