% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{initialAnalysis}
\alias{initialAnalysis}
\title{Perform basic analysis on the transposed object}
\usage{
initialAnalysis(f, npcs = 30, dims = 1:20)
}
\arguments{
\item{f}{\itemize{
\item A Seurat object
}}

\item{npcs}{\itemize{
\item Number of pcs to compute, defaults to 30
}}

\item{dims}{\itemize{
\item The dimensions to be used for UMAP and NN graph
}}
}
\value{
A Seurat object
}
\description{
This is typically called by findFPrime
}
