% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ligandReceptorResults}
\alias{ligandReceptorResults}
\title{ligandReceptorResults}
\format{
A list of data frames.

\describe{
A list containing:
interactionsOnEdges - a data frame whose first two columns give
the neighbouring cells and next two columns give their corresponding
clusters. Each of the remaining columns is a logical
corresponding to a ligand-receptor pair telling whether the ligand
is expressed in the first cell and the receptor is expressed in the
second cell.
totalInteractionsByCluster - a dataframe where the first column gives a
directed (sender-receiver) pair of clusters. The second column gives the
total number of edges between those clusters. The remaining columns give the
total numbers of edges on which particular ligand receptor interactions are
present.
meanInteractionsByCluster - a dataframe where the first column gives a
directed (sender-receiver) pair of clusters. The second column gives the
total number of edges between those clusters. The remaining columns give the
total numbers of edges on which particular ligand receptor interactions are
present (for that cluster pair) divided by the total number of edges between
those clusters.
simResults - a dataframe where the rownames are sender-receiver cluster pairs
and column names are ligand receptor pairs. Values give the number of
simulations for which observed values are greater than simulated values.
pValues - a dataframe where the rownames are sender-receiver cluster pairs
and column names are ligand receptor pairs. Entries are uppertail pvalues
describing whether a particular ligand receptor interaction is observed more
frequently between 2 clusters than expected.
}
}
\source{
Created from smallXenium and delaunayNeighbours by using
performLigandReceptorAnalysis(()
}
\usage{
ligandReceptorResults
}
\description{
The result of performLigandReceptorAnalysis(smallXenium, delaunayNeighbours,
"mouse", clusters,verbose=FALSE)
}
\keyword{datasets}
