% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleExampleServer.R
\name{make.getExample}
\alias{make.getExample}
\title{This function makes the function whichretrieves and
makes example data objects.}
\usage{
make.getExample()
}
\value{
This returns the function which retrieves
and makes example data objects. The latter saves any
object it has found for quicker return.  Using the value 'list'
causes it to return the list of all objects found so far.
}
\description{
This function makes the function whichretrieves and
makes example data objects.
}
\examples{
getExample = make.getExample()
## Provided:
smallXenium = getExample('smallXenium')
## Computed:
delaunayNeighbours = getExample('delaunayNeighbours')
}
