% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{makeFPrime}
\alias{makeFPrime}
\title{Create the transpose object of a given Seurat object}
\usage{
makeFPrime(f, active.assay, npcs = 30, dims = 1:20, res = 2)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object
}}

\item{active.assay}{\itemize{
\item active assay, defaults to the current active
assay
}}

\item{npcs}{\itemize{
\item number of pcs to find, defaults to 30
}}

\item{dims}{\itemize{
\item dimensions to be used for UMAP
}}

\item{res}{\itemize{
\item the resolution to be used for Louvain clustering
}}
}
\value{
The transposed Seurat object
}
\description{
This transposes the given Seurat object and performs basic
analysis including finding variable features, scaling data, running
PCA and UMAP, finding neighbors and clustering.
}
