% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{makeInteractionHeatmap}
\alias{makeInteractionHeatmap}
\title{This function takes interactionResults and plots a heatmap of -log10(pvalues).}
\usage{
makeInteractionHeatmap(
  interactionResults,
  clusters,
  colors = c(),
  pValCutoffClusterPair = 0.05,
  pValCutoffLigRec = 0.05,
  labelClusterPairs = T
)
}
\arguments{
\item{interactionResults}{\itemize{
\item as returned by performInteractionAnalysis()
}}

\item{clusters}{\itemize{
\item named vector of cell types where names are each cell and
clusters are a factor
}}

\item{colors}{\itemize{
\item a named list of colors where names are clusters. If not
specified the default pheatmap color scheme will be used.
}}

\item{pValCutoffClusterPair}{\itemize{
\item a cutoff for showing interactions between two
clusters. A cluster pair must have at least one ligand-receptor interaction
pvalue <  pValCutoffClusterPair. Defaults to 0.05.
}}

\item{pValCutoffLigRec}{\itemize{
\item a cutoff for showing interactions between a
ligand and receptor. At least one cluster pair must have
pvalue <  pValCutoffLigRec for ligand-receptor pair. Defaults to 0.05.
}}

\item{labelClusterPairs}{\itemize{
\item show labels for cluster pairs. Defaults to TRUE.
}}
}
\description{
This function takes interactionResults and plots a heatmap of -log10(pvalues).
}
