% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{makeSummedInteractionHeatmap}
\alias{makeSummedInteractionHeatmap}
\title{This function takes interactionResults and plots a heatmap of the total
number of ligand receptor interactions between clusters.}
\usage{
makeSummedInteractionHeatmap(interactionResults, clusters, type, logScale = T)
}
\arguments{
\item{interactionResults}{\itemize{
\item as returned by performInteractionAnalysis()
}}

\item{clusters}{\itemize{
\item named vector of cell types where names are each cell and
clusters are a factor
}}

\item{type}{\itemize{
\item "total" or "mean" to plot raw total interactions or mean interactions per edge.
}}

\item{logScale}{\itemize{
\item plot heatmap using log scale (defaults to T)
}}
}
\description{
This function takes interactionResults and plots a heatmap of the total
number of ligand receptor interactions between clusters.
}
