% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{meanGeneClusterOnCellUMAP}
\alias{meanGeneClusterOnCellUMAP}
\title{Mean gene cluster on cell umap}
\usage{
meanGeneClusterOnCellUMAP(f, fPrime, geneCluster)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object of cells
}}

\item{fPrime}{\itemize{
\item the corresponding Seurat object of genes
}}

\item{geneCluster}{\itemize{
\item a gene cluster of fPrime
}}
}
\value{
This returns a ggplot object
}
\description{
This function paints gene expression for a
given gene cluster on cell umap.
}
\examples{
g = meanGeneClusterOnCellUMAP(S,STranspose,geneCluster=0)
}
