% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moransI}
\alias{moransI}
\title{moransI}
\format{
A data fame containing Moran's I and related pvalues.

\describe{
Moran's I values calculated for the genes in smallXenium (using the SCT
assay). Pvalues derived using 100 permutations.
}
}
\source{
Created from smallXenium and delaunayNeighbours by using
runMoransI()
}
\usage{
moransI
}
\description{
A data fame containing Moran's I and related pvalues.
}
\keyword{datasets}
