% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moransILigandReceptor}
\alias{moransILigandReceptor}
\title{moransILigandReceptor}
\format{
A data frame showing the spatial autocorrelation of the
28 ligand receptor pairs

\describe{
A data frame with rownames giving the  28 ligand-receptor pairs and columns moransI
and pValues
}
}
\source{
Computed using the function runMoransI on the object edgeSeurat and
neighbours edgeNeighbours = computeEdgeGraph(delaunayNeighbours) with 100 trials.
For more informations see the CatsCradleSpatial vignette.
}
\usage{
moransILigandReceptor
}
\description{
Moran's I for the ligand receptor pairs
}
\keyword{datasets}
