% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseLNR}
\alias{mouseLNR}
\title{mouseLRN}
\format{
a data frame with two columns, 'from' and 'to'
\describe{A data frame with two columns, 'from' and 'to'.
Each row represents a mouse ligand - receptor pair.
}
}
\source{
This is derived from the nichenetr mouse ligand -
receptor network.
}
\usage{
mouseLNR
}
\description{
A data frame giving 11592 mouse ligand receptor pairs
}
\keyword{datasets}
