% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mouseLRN}
\alias{mouseLRN}
\title{mouseLRN}
\format{
a data frame with two columns, 'from' and 'to'
\describe{A data frame with two columns, 'from' and 'to'.
Each row represents a mouse ligand - receptor pair.
}
}
\source{
This is taken from the nichenetr package,
url = {https://www.nature.com/articles/s41592-019-0667-5}.
Specifically, we use the mouse ligand - receptor network.
}
\usage{
mouseLRN
}
\description{
A data frame giving 11592 mouse ligand receptor pairs
}
\keyword{datasets}
