% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBHDByCTMatrixExtended}
\alias{NBHDByCTMatrixExtended}
\title{NBHDByCTMatrixExtended}
\format{
A matrix whose rows correspond to extended neighbourhoods
(and therefore, to cells) and whose columns correspond to
cell types.

\describe{
This matrix gives the counts for each of the cell types
in each extended neighbourhood.
}
}
\source{
This is computed from delaunayNeighbours and
clusters using computeNBHDByCTMatrix().
}
\usage{
NBHDByCTMatrixExtended
}
\description{
A neighbourhoods by cell types matrix giving the count
of cells of each type in each extended neighbourhood.
}
\keyword{datasets}
