% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBHDByCTSeurat}
\alias{NBHDByCTSeurat}
\title{NBHDByCTSeurat}
\format{
A Seurat object consisting of 4261 samples (the
neighbourhoods and 24 features (the cell types).

\describe{
This is a Seurat object created by taking NBHDByCTMatrix
as the counts.
}
}
\source{
Created from NBHDByCTMatrix by
computeNBHDVsCTSeurat()
}
\usage{
NBHDByCTSeurat
}
\description{
A Seurat object computed from NBHDByCTMatrix. Think of
neighbourhoods "expressing" cell types.
}
\keyword{datasets}
