% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NBHDByCTSeuratExtended}
\alias{NBHDByCTSeuratExtended}
\title{NBHDByCTSeuratExtended}
\format{
A Seurat object consisting of 4261 samples (the
neighbourhoods) and 24 features (the cell types).

\describe{
This is a Seurat object created by taking NBHDByCTMatrixExtended
as the counts.
}
}
\source{
Created from NBHDByCTMatrixExtended by
computeNBHDVsCTSeurat()
}
\usage{
NBHDByCTSeuratExtended
}
\description{
A Seurat object computed from NBHDByCTMatrixExtended. Think of
neighbourhoods "expressing" cell types.
}
\keyword{datasets}
