% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{nearbyGenes}
\alias{nearbyGenes}
\title{Nearby genes}
\usage{
nearbyGenes(fPrime, geneSet, radius, metric = "umap", numPCs = NULL)
}
\arguments{
\item{fPrime}{\itemize{
\item a Seurat object of genes
}}

\item{geneSet}{\itemize{
\item set of genes
}}

\item{radius}{\itemize{
\item the distance around the given set
}}

\item{metric}{\itemize{
\item the metric to use, one of umap, tsne,
pca or nearest neighbor
}}

\item{numPCs}{\itemize{
\item used only if the metric is pca
}}
}
\description{
This finds the genes near a give subset using either
a dimensional reduction or the nearest neighbor graph
}
\examples{
geneSet = intersect(colnames(STranspose),hallmark[[1]])
geometricallyNearby = nearbyGenes(STranspose,geneSet,radius=0.2,metric='umap')
combinatoriallyNearby = nearbyGenes(STranspose,geneSet,radius=1,metric='NN')
}
