% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{neighborTerms}
\alias{neighborTerms}
\title{This function predicts the functions of genes based on the functions of their
neighbours.}
\usage{
neighborTerms(fPrime, genesAnno, normalise = T)
}
\arguments{
\item{fPrime}{\itemize{
\item a transposed Seurat object (generated with
transposeSeuratObject())
}}

\item{genesAnno}{\itemize{
\item genes annotated with gene sets
}}
}
\value{
\itemize{
\item A list where names are genes and values are lists of terms.
}
}
\description{
This function predicts the functions of genes based on the functions of their
neighbours.
}
