% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NN}
\alias{NN}
\title{NN}
\format{
A data frame with columns nodeA, nodeB and weight
\describe{
This gives the weigthed edges of the nearest neighbor graph
of the genes in STranspose.
}
}
\source{
Extracted from STranspose by getNearestNeighborListsSeurat
}
\usage{
NN
}
\description{
The nearest neighbor graph of STranspose
}
\keyword{datasets}
