% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{performInteractionAnalysis}
\alias{performInteractionAnalysis}
\title{Given a seurat object, a spatial graph, clusters and species this function
identifies ligand-receptor interactions between neighbouring cells,
identifies ligand-receptor interactions within and between clusters and
calculates whether these are observed more frequently than expected by
chance.}
\usage{
performInteractionAnalysis(
  obj,
  spatialGraph,
  species,
  clusters,
  nSim = 1000,
  lrn = getLigandReceptorNetwork(species),
  verbose = TRUE
)
}
\arguments{
\item{obj}{\itemize{
\item a Seurat object
}}

\item{spatialGraph}{\itemize{
\item a data frame of neighbouring
cell pairs.
}}

\item{species}{\itemize{
\item either 'human' or 'mouse'
}}

\item{clusters}{\itemize{
\item named vector of clusters where names are each cell and
clusters are a factor
}}

\item{nSim}{\itemize{
\item number of simulations to perform for p value calculation.
}}

\item{lrn}{\itemize{
\item a ligand-receptor network, i.e., a
data frame with columns from and to.  By default, it
retrieves the nichenetr ligand receptor network
}}

\item{verbose}{\itemize{
\item whether to print trace, defaults to TRUE
}}
}
\value{
This returns a data frame whose first two columns give
the neighbouring cells.  Each of the remaining columns is a logical
corresponding to a ligand-receptor pair telling whether the ligand
is expressed in the first cell and the receptor is expressed in the
second cell.
}
\description{
Given a seurat object, a spatial graph, clusters and species this function
identifies ligand-receptor interactions between neighbouring cells,
identifies ligand-receptor interactions within and between clusters and
calculates whether these are observed more frequently than expected by
chance.
}
