% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleLigandReceptorAnalysis.R
\name{permuteColumns}
\alias{permuteColumns}
\title{This function permutes the columns of a sparse dcG matrix.}
\usage{
permuteColumns(M)
}
\arguments{
\item{M}{\itemize{
\item a binarised expression matrix in sparse format where rows are cells and columns
}}
}
\value{
This returns a matrix in which the values have been permuted within
columns.
}
\description{
This function permutes the columns of a sparse dcG matrix.
}
