% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{permuteMatrix}
\alias{permuteMatrix}
\title{This function permutes the rows of a matrix.}
\usage{
permuteMatrix(M)
}
\arguments{
\item{M}{\itemize{
\item a binarised expression matrix where rows are genes and columns
}}
}
\value{
This returns a matrix in which the values have been permuted within
rows.
}
\description{
This function permutes the rows of a matrix.
}
