% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{predictTerms}
\alias{predictTerms}
\title{This function makes annotation predictions for a
set of genes based on gene sets (e.g., hallmark)
and a CatsCradle object by considering the annotations of
its neighboring genes.}
\usage{
predictTerms(
  genes,
  geneSets,
  fPrime,
  normaliseByGeneSet = TRUE,
  normaliseByEdgeWeight = TRUE,
  normaliseToUnitVector = TRUE
)
}
\arguments{
\item{genes}{\itemize{
\item a character vector of genes
}}

\item{geneSets}{\itemize{
\item a set of annotations, e.g., hallmark
or GO
}}

\item{fPrime}{\itemize{
\item a Seurat object of genes
}}

\item{normaliseByGeneSet}{\itemize{
\item determines whether vector annotations
are normalised by gene set size.  Defaults to TRUE
}}

\item{normaliseByEdgeWeight}{\itemize{
\item determines whether neighbor
contributions are normalised by edge weight.  Defaults to
TRUE.
}}

\item{normaliseToUnitVector}{\itemize{
\item determines whether to normalise
returned values to unit length.  Defaults to TRUE
}}
}
\description{
This function makes annotation predictions for a
set of genes based on gene sets (e.g., hallmark)
and a CatsCradle object by considering the annotations of
its neighboring genes.
}
\examples{
set.seed(100)
genes = sample(colnames(STranspose),5)
predictions = predictTerms(genes,hallmark,STranspose)

}
