% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{predictTermsAllGenes}
\alias{predictTermsAllGenes}
\title{This function predicts the functions of all genes based
on the functions of their neighbours.}
\usage{
predictTermsAllGenes(
  fPrime,
  genesAnno,
  genes = colnames(fPrime),
  normalise = T
)
}
\arguments{
\item{fPrime}{\itemize{
\item a transposed Seurat object (generated with
transposeSeuratObject())
}}

\item{genesAnno}{\itemize{
\item genes annotated with gene sets
}}

\item{genes}{\itemize{
\item the genes to be predicted, defaults to all
}}

\item{normalise}{\itemize{
\item whether to normalise to gene set size, defauls
to TRUE
}}
}
\value{
\itemize{
\item A list where names are genes and values are lists of terms.
The values of the lists of terms are calculated according to the weights
of the edges connecting the neighbors.
}
}
\description{
This function predicts the functions of all genes based
on the functions of their neighbours.
}
\examples{
set.seed(100)
genes = sample(colnames(STranspose),50)
genesAnno = annotateGenes(hallmark)
predictions = predictTermsAllGenes(STranspose,genesAnno,genes)
}
