% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleNNGraphs.R
\name{randomiseGraph}
\alias{randomiseGraph}
\title{This function performs degree-preserving randomisation of neighbour graphs.}
\usage{
randomiseGraph(spatialGraph, maxTries = 1000)
}
\arguments{
\item{spatialGraph}{\itemize{
\item a spatial graph in neighbour list format.
}}

\item{maxTries}{\itemize{
\item the maximum number of tries to remove self edges during
graph randomisation. If self edges are remeining this will be reported.
}}
}
\value{
A randomised graph where degree from the original graph is preserved.
}
\description{
This function performs degree-preserving randomisation of neighbour graphs.
}
