% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{randomiseNodeIndices}
\alias{randomiseNodeIndices}
\title{This function generates random indices for node B}
\usage{
randomiseNodeIndices(neighborListDf, n = 100, useWeights = F)
}
\arguments{
\item{neighborListDf}{\itemize{
\item a dataframe containing the neighborlist
}}

\item{n}{\itemize{
\item the number of times to randomise indices
}}

\item{useWeights}{\itemize{
\item whether to preserve edgeweights.
}}
}
\value{
\itemize{
\item a matrix with randomised indices for node B
}
}
\description{
This function generates random indices for node B
}
\examples{
NN = desymmetriseNN(NN)
randomIndices = randomiseNodeIndices(NN,10,TRUE)
}
