% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{reduceCombinatorialBalls}
\alias{reduceCombinatorialBalls}
\title{This function reduces combinatorial balls to an
extended nearest neighbour graph}
\usage{
reduceCombinatorialBalls(balls)
}
\arguments{
\item{balls}{\itemize{
\item A named list as produced by the function
findCombinatorialNeighbourhoods
}}
}
\value{
This returns a nearest neighbour data frame where the cells
in each combinatorial ball are now considered the center's
neighbours.
}
\description{
This function reduces combinatorial balls to an
extended nearest neighbour graph
}
