% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S}
\alias{S}
\title{S}
\format{
A Seurat object

\describe{
A Seurat object of cells.  It includes a UMAP of the
cells and annotated clustering into cell types.
}
}
\source{
This is subset from the data associated with
https://www.nature.com/articles/s41586-021-04006-z
}
\usage{
S
}
\description{
A Seurat object of 2000 genes by 1445 cells.
}
\keyword{datasets}
