% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{S_sce}
\alias{S_sce}
\title{S_sce}
\format{
A SingleCellExperiment object

\describe{
A SingleCellExperiment version of S.
}
}
\source{
This was made from S using as.SingleCellExperiment
}
\usage{
S_sce
}
\description{
A SingleCellExperiment object of 2000 genes by 1445 cells.
}
\keyword{datasets}
