% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{sankeyFromMatrix}
\alias{sankeyFromMatrix}
\title{This makes a sankey graph from a matrix of average
express.  Our "Cat's Cradle".}
\usage{
sankeyFromMatrix(
  M,
  disambiguation = c("R_", "C_"),
  fontSize = 20,
  minus = "red",
  plus = "blue"
)
}
\arguments{
\item{M}{\itemize{
\item a matrix of gene expression
}}

\item{disambiguation}{\itemize{
\item used to distinguish between
the row names and the column names if these overlap
}}

\item{fontSize}{\itemize{
\item defaults to 20
}}

\item{minus}{\itemize{
\item color to use for links with negative
values
}}

\item{plus}{\itemize{
\item color for positive values
}}
}
\value{
A sankey graph
}
\description{
This makes a sankey graph from a matrix of average
express.  Our "Cat's Cradle".
}
