% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratCells}
\alias{seuratCells}
\title{seuratCells}
\format{
A vector of cells

\describe{
A vector of cells consisting of half the cells from
each seurat_cluster in exSeuratObj used to subset this
object to give toy examples.
}
}
\source{
Computed by retrieving half the cells from each
cluster in exSeuratObj
}
\usage{
seuratCells
}
\description{
A vector of cells used for subsetting exSeuratObj
}
\keyword{datasets}
