% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuratGenes}
\alias{seuratGenes}
\title{seuratGenes}
\format{
A vector of genes

\describe{
A vector of the top 100 most variable genes in exSeuratObj
used to subset this object to give toy examples.
}
}
\source{
Computed by retrieving the data layer from exSeuratObj and
subsetting to the 100 genes with the highest standard deviation.
}
\usage{
seuratGenes
}
\description{
A vector of genes used for subsetting exSeuratObj
}
\keyword{datasets}
