% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallCellTypesPerCellTypeMatrix}
\alias{smallCellTypesPerCellTypeMatrix}
\title{smallCellTypesPerCellTypeMatrix}
\format{
A matrix whose rows and columns correspond to
cell types.

\describe{
Each row of this matrix corresponds to a cell type.  On
that row we see the proportions of all neighbourhoods
surrounding cells of that cell type as regards the cell types
they contain.  In particular, each row sums to 1.  Since
this comes from a subset of the full Xenium object, not all
cell types appear.
}
}
\source{
This is created from smallNbhdMatrix and the seurat_clusters
of smallXenium using cellTypesPerCellTypeMatrix()
}
\usage{
smallCellTypesPerCellTypeMatrix
}
\description{
For each cell type, this matrix shows the fraction
of the neighbourhoods of that cell type composed of
each cell type.
}
\keyword{datasets}
