% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallCentroids}
\alias{smallCentroids}
\title{smallCentroids}
\format{
A data frame with columns x, y, and cell.  Rownames
also give the names of the cells.

\describe{
This data frame gives the centroids of the cells in smallXenium.
}
}
\source{
Extracted from smallXenium with GetTissueCoordinates().
The names of the cells are appended as rownames.
}
\usage{
smallCentroids
}
\description{
A data.frame of the centroids of the cells in the
smallXenium Seurat object
}
\keyword{datasets}
