% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallCombNbhds}
\alias{smallCombNbhds}
\title{smallCombNbhds}
\format{
A named list.  Each entry in this lists gives
the names of the cells in the combinatorial ball of
radius 2 surrounding a given cell.

\describe{
A named list.  This gives the cells in the combinatorial
ball of radius 2 around each cell.
}
}
\source{
This is computed from smallDelaunayTriangulation using
findCombinatorialNeighbourhoods() with radius 2
}
\usage{
smallCombNbhds
}
\description{
A named list giving the combinatorial neighbourhood
of radius 2 around each cell
}
\keyword{datasets}
