% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallDelaunayTriangulation}
\alias{smallDelaunayTriangulation}
\title{smallDelaunayTriangulation}
\format{
A data frame with character columns nodeA and nodeB.

\describe{
The entries in this data frame are cell names from smallXenium.
Each row represents an undirected edge in the Delaunay
triangulation of smallCentroids.
}
}
\source{
This is computed from smallCentroids using
computeNeighboursDelaunay()
}
\usage{
smallDelaunayTriangulation
}
\description{
A data frame of the Delaunay triangulation of smallCentroids,
the centroids of sammXenium
}
\keyword{datasets}
