% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallNbhdMatrix}
\alias{smallNbhdMatrix}
\title{smallNbhdMatrix}
\format{
A matrix whose rows correspond to neighbourhoods
(and therefore, to cells) and whose columns correspond to
cell types.

\describe{
This matrix gives the counts for each of the cell types
in each neighbourhoods.  Since it is taken from the subset
smallXenium rather from the full Xenium object, not all
cell types appear.
}
}
\source{
This is computed from smallNbhds and
smallXenium$seurat_clusters using computeNeighbourhoodByCTMatrix().
}
\usage{
smallNbhdMatrix
}
\description{
A neighbourhoods by cell types matrix giving the count
of cells of each type in each neighbourhood
}
\keyword{datasets}
