% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallNbhdObj}
\alias{smallNbhdObj}
\title{smallNbhdObj}
\format{
A Seurat object consisting of 4261 samples (the
neighbourhoods and 24 features (the cell types).

\describe{
This is a Seurat object created by taking smallNbhdMatrix
as the counts.
}
}
\source{
Created from smallNbhdMatrix by
computeNeighbourhoodByCTSeurat()
}
\usage{
smallNbhdObj
}
\description{
A Seurat object computed from smallNbhdMatrix. Think of
neighbourhoods "expressing" cell types.
}
\keyword{datasets}
