% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallNbhds}
\alias{smallNbhds}
\title{smallNbhds}
\format{
A named list.  Each entry in this lists gives
the names of the cells surrounding a given cell.

\describe{
This list gives the combinatorial balls of radius 1
around the cells in smallDelaunayTriangulation
}
}
\source{
Computed from smallDelaunayTriangulation using
computeNeighbourhoods() with addSelf=TRUE
}
\usage{
smallNbhds
}
\description{
A named list of characters in which each name is the
name of a cell from smallXenium and its vector gives
the names of the surrounding cells.
}
\keyword{datasets}
