% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallXenium}
\alias{smallXenium}
\title{samllXenium}
\format{
A Seurat object

\describe{
A spatial Seurat object subset from the Xenium object used in
https://satijalab.org/seurat/articles/seurat5_spatial_vignette_2
Other data objects whose names begin with small are derived from
this object using CatsCradle.
}
}
\source{
This is subset from the Xenium spatial Seurat object
https://cf.10xgenomics.com/samples/xenium/1.0.2/Xenium_V1_FF_Mouse_Brain_Coronal_Subset_CTX_HP/Xenium_V1_FF_Mouse_Brain_Coronal_Subset_CTX_HP_outs.zip
to include a small region of the field of view surrounding the dentate gyrus.
}
\usage{
smallXenium
}
\description{
A spatial Seurat object of 4261 cells and 248 genes
}
\keyword{datasets}
