% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STranspose}
\alias{STranspose}
\title{STranspose}
\format{
A Seurat object
\describe{
This is the transpose of S. It includes a UMAP of the
genes and and their clustering into gene types.
}
}
\source{
Produced from S by transposeSeuratObject()
}
\usage{
STranspose
}
\description{
A Seurat object of 1445 cells by 2000 genes
}
\keyword{datasets}
