% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STranspose_sce}
\alias{STranspose_sce}
\title{STranspose_sce}
\format{
A SingleCellExperiment object
\describe{
This is a SingleCellExperiment version of STranspose.
}
}
\source{
Produced from STranspose using as.SingleCellExperiment
together with ancillary code for copying nearest neighbour
graphs
}
\usage{
STranspose_sce
}
\description{
A SingleCellExperiment object of 1445 cells by 2000 genes
}
\keyword{datasets}
