% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{stripGeneSet}
\alias{stripGeneSet}
\title{This function strips out non-gene information from
the beginning of GO sets, etc.}
\usage{
stripGeneSet(geneSet)
}
\arguments{
\item{geneSet}{\itemize{
\item a list of gene sets
}}
}
\value{
a named list of gene sets
}
\description{
This function strips out non-gene information from
the beginning of GO sets, etc.
}
