% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{symmetriseNN}
\alias{symmetriseNN}
\title{This symmetrises a nearest neighbors graph.}
\usage{
symmetriseNN(NN)
}
\arguments{
\item{NN}{\itemize{
\item a nearest neighbors graph as returned
by getNearestNeighbourLists
}}
}
\value{
a nearest neighbors graph
}
\description{
This first checks to see if the NN graph is symmetric
and if not symmetrises it.
}
\examples{
NN = make.getExample()('NN')
NNStar = symmetriseNN(NN)
}
