% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{symmetrizeNN}
\alias{symmetrizeNN}
\title{This symmetrizes a nearest neighbors graph.}
\usage{
symmetrizeNN(NN)
}
\arguments{
\item{NN}{\itemize{
\item a nearest neighbors graph as returned
by getNearestNeighborListsSeurat
}}
}
\value{
a nearest neighbors graph
}
\description{
This first checks to see if the NN graph is symmetric
and if not symmetrizes it.
}
