% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{tagRowAndColNames}
\alias{tagRowAndColNames}
\title{This gussies up the rownames and colnames of M}
\usage{
tagRowAndColNames(M, ccTag = "CC_", gcTag = "GC_")
}
\arguments{
\item{M}{\itemize{
\item a matrix, typically the average expression matrix
}}

\item{ccTag}{\itemize{
\item a prefix for the row (cell cluster) names
}}

\item{gcTag}{\itemize{
\item a prefix for the column (gene cluster) names
}}
}
\description{
This gussies up the rownames and colnames of M
}
