% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{transposeSeuratObject}
\alias{transposeSeuratObject}
\title{Create the transpose of a Seurat object}
\usage{
transposeSeuratObject(f, active.assay = f@active.assay)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object
}}

\item{active.assay}{\itemize{
\item the assay to use.  Defaults to the
active.assay
}}
}
\value{
A Seurat object
}
\description{
This takes a Seurat object f and creates a new Seurat object whose
expression matrix is the transpose of that of f. Should preferably
be called through makeFPrime()
}
