% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xeniumCells}
\alias{xeniumCells}
\title{xeniumCells}
\format{
A vector of cells

\describe{
A vector of cells consisting of approximately one
quarter of the cells in smallXenium used to subset
this object to give toy examples.
}
}
\source{
We extracted a rectangle whose width and
height were one half the width and height of smallXenium
and which was centered in the field of view of
smallXenium
}
\usage{
xeniumCells
}
\description{
A vector of cells used for subsetting exSeuratObj
}
\keyword{datasets}
