% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{aggregateGeneExpression}
\alias{aggregateGeneExpression}
\title{This function takes a Seurat object and a list of
neighbourhoods and creates a Seurat object where the
columns are the neighbourhoods, the rows are are the
genes and the values are gene expression totals for
the cells in each neighbourhood}
\usage{
aggregateGeneExpression(f, neighbourhoods)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object with layer counts
}}

\item{neighbourhoods}{\itemize{
\item Neighbourhoods as given by a
collapsed expanded edge graph, as produced by
collapseNeighbourhoods. In particular, each cell should
appear as nodeA.
}}
}
\value{
a Seurat object giving total gene expression
in each neighbourhood.
}
\description{
This function takes a Seurat object and a list of
neighbourhoods and creates a Seurat object where the
columns are the neighbourhoods, the rows are are the
genes and the values are gene expression totals for
the cells in each neighbourhood
}
