% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{annotateGenes}
\alias{annotateGenes}
\title{This function annotates genes with terms}
\usage{
annotateGenes(geneSets)
}
\arguments{
\item{geneSets}{\itemize{
\item a list of gene sets, e.g., as produced by
readGmt
}}
}
\value{
\itemize{
\item A list where names are genes and values are lists of terms
}
}
\description{
This function annotates genes with terms
}
