% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeNeighbourhoodByCTSeurat}
\alias{computeNeighbourhoodByCTSeurat}
\title{This function creates a seurat object using a neighbourhood by cell type
matrix}
\usage{
computeNeighbourhoodByCTSeurat(
  neighbourhoodByCT,
  resolution = 0.1,
  npcs = 10,
  n.neighbors = 30L,
  compute_UMAP = T,
  transpose = F
)
}
\arguments{
\item{neighbourhoodByCT}{\itemize{
\item a matrix of neighbourhoods by cell types
}}

\item{resolution}{\itemize{
\item resolution for clustering (default 0.1)
}}

\item{npcs}{\itemize{
\item number of pcs used for PCA, defaults to 10
}}

\item{n.neighbors}{\itemize{
\item number of neighbors used by UMAP, defaults to 30
}}

\item{compute_UMAP}{\itemize{
\item defaults to TRUE
}}

\item{transpose}{\itemize{
\item defaults to FALSE
}}
}
\value{
a seurat object based on a neighbourhood by cell type matrix,
containing clusters and UMAP.
}
\description{
This function creates a seurat object using a neighbourhood by cell type
matrix
}
\examples{
smallNbhdObj = computeNeighbourhoodByCTSeurat(smallNbhdMatrix)
}
