% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{computeNeighbourhoods}
\alias{computeNeighbourhoods}
\title{This function extracts neighbourhoods from a spatial graph in neighbour list
format}
\usage{
computeNeighbourhoods(
  spatialGraph,
  cellNames = extractCells(spatialGraph),
  addSelf = F
)
}
\arguments{
\item{spatialGraph}{\itemize{
\item a spatial graph in neighbour list format.
}}
}
\value{
a named list of neighbourhoods where a neighborhood is a set of a
cell's nearest neighbours.
}
\description{
This function extracts neighbourhoods from a spatial graph in neighbour list
format
}
