% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleExampleServer.R
\name{exampleObjects}
\alias{exampleObjects}
\title{This returns the names of available example objects.}
\usage{
exampleObjects()

exampleObjects()
}
\value{
A character vector of the names of available
example data objects

A character vector of the names of the available
example objects
}
\description{
This returns the names of available example objects.

A listing of the names of the available example objects.
}
\examples{
availableObjects = exampleObjects()
availableObjects = exampleObjects()
}
