% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{getAverageExpressionMatrix}
\alias{getAverageExpressionMatrix}
\title{This computes average expression of each gene cluster in
each cell cluster and returns the result as a matrix}
\usage{
getAverageExpressionMatrix(f, fPrime)
}
\arguments{
\item{f}{\itemize{
\item The Seurat object of cells
}}

\item{fPrime}{\itemize{
\item The Seurat object of genes
}}
}
\value{
A matrix of the average expression
}
\description{
This computes average expression of each gene cluster in
each cell cluster and returns the result as a matrix
}
\examples{
M = getAverageExpressionMatrix(S,STranspose)
}
