% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{getClusterOrder}
\alias{getClusterOrder}
\title{This gets the clusters in their cannonical order}
\usage{
getClusterOrder(f)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object with meta.data column seurat_clusters
}}
}
\value{
A vector of these unique values in order
}
\description{
Note: seurat_clusters might be either a factor or a numeric.
This will keep the class.
}
