% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{getNearestNeighborListsSeurat}
\alias{getNearestNeighborListsSeurat}
\title{This function extracts a shared nearest neighbor network
from a Seurat object}
\usage{
getNearestNeighborListsSeurat(f, assay = "RNA")
}
\arguments{
\item{f}{\itemize{
\item a Seurat object
}}

\item{assay}{\itemize{
\item name of assay used to construct the SNN graph
(defaults to "RNA")
}}
}
\value{
\itemize{
\item This returns dataframe of neighbors:
nodeA - node names for node A
nodeB - node names for node B
weight - edge weight
}
}
\description{
This function extracts a shared nearest neighbor network
from a Seurat object
}
