% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradleSpatial.R
\name{makeSummedLRInteractionHeatmap}
\alias{makeSummedLRInteractionHeatmap}
\title{This function takes ligandReceptorResults and plots a heatmap of the total
number of ligand receptor interactions between clusters.}
\usage{
makeSummedLRInteractionHeatmap(
  ligandReceptorResults,
  clusters,
  type,
  logScale = TRUE
)
}
\arguments{
\item{ligandReceptorResults}{\itemize{
\item as returned by performLigandReceptorAnalysis()
}}

\item{clusters}{\itemize{
\item named vector of cell types where names are each cell and
clusters are a factor
}}

\item{type}{\itemize{
\item "total" or "mean" to plot raw total interactions or mean interactions per edge.
}}

\item{logScale}{\itemize{
\item plot heatmap using log scale (defaults to T)
}}
}
\value{
matrix of total ligand receptor interactions that underlies the heatmap.
}
\description{
This function takes ligandReceptorResults and plots a heatmap of the total
number of ligand receptor interactions between clusters.
}
\examples{
cellTypePerCellTypeLigRecMatrix = 
makeSummedLRInteractionHeatmap(ligandReceptorResults, clusters, "mean")
}
