% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moransILigandReceptor}
\alias{moransILigandReceptor}
\title{moransILigandReceptor}
\format{
A data fame containing Moran's I for ligand receptor pairs and
related pvalues.

\describe{
Moran's I values calculated for the ligand receptor pairs in edgeSeurat.
Pvalues derived using 100 permutations.
}

A data frame showing the spatial autocorrelation of the
28 ligand receptor pairs

\describe{
A data frame with rownames giving the  28 ligand-receptor pairs and columns moransI
and pValues
}
}
\source{
Created from edgeSeurat and edgeNeighbours by using runMoransI()

Computed using the function runMoransI on the object edgeSeurat and
neighbours edgeNeighbours = computeEdgeGraph(delaunayNeighbours) with 100 trials.
For more informations see the CatsCradleSpatial vignette.
}
\usage{
moransILigandReceptor

moransILigandReceptor
}
\description{
A data fame containing Moran's I for the spatial distribution of the presence
of ligand receptor pairs on edges.

Moran's I for ligand receptor pairs
}
\keyword{datasets}
