% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moransILigandReceptor}
\alias{moransILigandReceptor}
\title{moransILigandReceptor}
\format{
A data fame containing Moran's I for ligand receptor pairs and
related pvalues.

\describe{
Moran's I values calculated for the ligand receptor pairs in edgeSeurat.
Pvalues derived using 100 permutations.
}
}
\source{
Created from edgeSeurat and edgeNeighbours by using runMoransI()
}
\usage{
moransILigandReceptor
}
\description{
A data fame containing Moran's I for the spatial distribution of the presence
of ligand receptor pairs on edges.
}
\keyword{datasets}
