% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{predictTerms}
\alias{predictTerms}
\title{This function predicts the functions of a gene based on the functions of its
neighbours.}
\usage{
predictTerms(
  gene,
  fPrime,
  genesAnno,
  normalise = T,
  graph = defaultGraph(fPrime)
)
}
\arguments{
\item{gene}{\itemize{
\item gene to annotate
}}

\item{fPrime}{\itemize{
\item a transposed Seurat object (generated with
transposeSeuratObject())
}}

\item{genesAnno}{\itemize{
\item genes annotated with gene sets
}}

\item{normalise}{\itemize{
\item choose whether to normalise
(divide scores by total weight of edges)
}}

\item{graph}{\itemize{
\item which graph to use.  Defaults to the
active.assay followed by _snn
}}
}
\value{
\itemize{
\item A lists of terms where values are scores.
Scores are calculated according to the weights of the SNN graph.
}
}
\description{
This function predicts the functions of a gene based on the functions of its
neighbours.
}
