% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{readGmt}
\alias{readGmt}
\title{This function reads in gene sets in .gmt format}
\usage{
readGmt(gmtFile)
}
\arguments{
\item{gmtFile}{\itemize{
\item a .gmt file containing gene sets
}}
}
\value{
\itemize{
\item A named list of gene sets
}
}
\description{
This function reads in gene sets in .gmt format
}
