% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shorterHallmark}
\alias{shorterHallmark}
\title{shorterHallmark}
\format{
A named list
\describe{
The alphabetically first ten gene hallmark gene sets.
}
}
\source{
Abbreviated from hallmark
}
\usage{
shorterHallmark
}
\description{
An abbreviation of mouse hallmark gene sets.
}
\keyword{datasets}
