% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{symmetryCheckNN}
\alias{symmetryCheckNN}
\title{Tests whether a nearest neighbor graph is symmetric}
\usage{
symmetryCheckNN(NN)
}
\arguments{
\item{NN}{\itemize{
\item a nearest neighbor graph.  This is in the form
of a data frame  as returned by getNearestNeighborListsSeurat.
Its coloumns include nodeA and nodeB.
}}
}
\value{
TRUE or FALSE
}
\description{
The nearest neighbor relationship is not inherently
symmetric.  This tests whether the nearest neighbor graph
retrieved from a Seurat object is.
}
