% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CatsCradle.R
\name{transposeSeuratObject}
\alias{transposeSeuratObject}
\title{Create the transpose of a Seurat object}
\usage{
transposeSeuratObject(f, active.assay = "RNA", npcs = 30, dims = 1:20, res = 1)
}
\arguments{
\item{f}{\itemize{
\item a Seurat object
}}

\item{active.assay}{\itemize{
\item the assay to use.  Defaults to 'RNA'
}}

\item{npcs}{\itemize{
\item number of principal components, defaults to 30
}}

\item{dims}{\itemize{
\item dimensions to use for umap and nearest neighbors,
defaults to 1:20
}}

\item{res}{\itemize{
\item the clustering resolution, defaults to 1
}}
}
\value{
A Seurat object
}
\description{
This takes a Seurat object f and creates a new Seurat object whose
expression matrix is the transpose of that of f.
}
\examples{
STranspose = transposeSeuratObject(S)
}
